Computer Version
An Eclipse Project. Currently an android app in the making, with
multiple game modes and a working laser upgrade system. Much of the game
logic code remains the same, however other code regarding the java.swing
library is completely different in the android version. This repository
is simply to give an idea of what the app intends to become.